refactor: 사용하지 않는 Repository 정리
